Refactored test framework code
DAT-8752
